day 4
explosion effects
death explosion
spawner random
charger ai
spread random targeting

+++ src/components/spawner.rs
@@ -8,11 +8,11 @@ pub struct Spawner {
 	pub spawn_rate: Timer,
 	//Time between spawning individual enemies
 	pub spawn_speed: Timer,
-	pub min_batch_size: usize,
-	pub max_batch_size: usize,
+	pub min_batch_size: u32,
+	pub max_batch_size: u32,
 	pub prefabs: Vec<Entity>,
 	pub spawn_effect: Entity,
 }
 
 #[derive(Component, Default, Reflect)]
-pub struct SpawnBatch(pub usize);
+pub struct SpawnBatch(pub u32);
